use valid swagger integer type

+++ app/controllers/api/v1/items_controller.rb
@@ -48,7 +48,7 @@ module Api::V1
           key :in, :path
           key :description, "item id"
           key :required, true
-          key :type, :int64
+          key :type, :integer
         end
         parameter do
           key :name, :q
